pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"


[project]
name = "sad2_final_project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.14.1"
dependencies = [
    "boolean-py>=5.0",
    "matplotlib>=3.10.8",
    "networkx>=3.6.1",
    "numpy>=2.4.1",
    "pandas>=2.3.3",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
